fix(gp): address PR #120 review comments
Predict path:
- Fix predict_mean centering bug (subtract prior mean from y_tilde)
- Share a single _kernel_context across (mean, var) so Pattern B/C
  prior'd kernels resample once per predict call
- Use kernel.diag(X_star) instead of an N-call vmap
- Route the predict-time Cholesky through _stable_cholesky so prediction
  matches the inference-time near-singular guard
- Drop unused _solver / lineax / cholesky imports; tighten the
  NonGaussConditionedGP docstring to match actual O(N^3)-per-call cost

Public API typing:
- Tighten condition_nongauss to (Likelihood, _NonGaussStrategy)
  -> NonGaussConditionedGP via TYPE_CHECKING
- Annotate SoftmaxLikelihood.log_prob's y as Int[Array, " N"]

Numerics:
- Switch _stable_cholesky's NaN fallback to lax.cond so the jittered
  factorization isn't computed in the common (PD) case
- Compute EP tilted moments in log-space with per-site logsumexp
  shifts, eliminating the overflow/underflow risk on heavy-tailed or
  large-logit likelihoods

+++ src/pyrox/gp/_likelihoods.py
@@ -33,7 +33,7 @@ from typing import Any
 import equinox as eqx
 import jax.numpy as jnp
 import numpyro.distributions as nd
-from jaxtyping import Array, Float
+from jaxtyping import Array, Float, Int
 
 from pyrox.gp._protocols import Likelihood
 
@@ -183,7 +183,7 @@ class SoftmaxLikelihood(Likelihood):
     def log_prob(
         self,
         f: Float[Array, "N C"],
-        y: Float[Array, " N"],
+        y: Int[Array, " N"],
     ) -> Float[Array, ""]:
         return nd.Categorical(logits=f).log_prob(y).sum()
